// front/src/app/models/projet.ts
import { Utilisateur } from "./utilisateur";

export class Projet {
    projet_id: number = 0;
    nom: string= '';
    description: string= '';
    prefixe: string= '';
    statut: string= '';
    isGestionnaireAnomalie: boolean=false;
    nomGestionnaire: string='';
    creatorId : number = 0;

  
      
    
    }